requirements.txt
flask==3.0.0
mysql-connector-python==9.0.0
flask-sqlalchemy==2.5.1
flask_wtf==1.2.2
flask-bcrypt==1.0.1
python-dotenv